Answer of the Questions
Q2. write your own general definitions of the meaning of "observations" and "variables" based on asking the Chatbot to explain these terms in the context of your dataset

Q4. If the dataset you're using has non-numeric variables and missing values in numeric variables, explain the discrepancies between size of the dataset given by df. shape and what is reported by df.describe() with respect to the number of columns it analyzes and the values if reports in the "count" column

Q5. Use your Chatbot session to help understand the difference between the following and then provide your own paraphrasing summarization of that difference